Helper namespace renamed to services (#8)
