+++ GitWorkTree/ViewModel/RelayCommand.cs
@@ -1,4 +1,4 @@
-using GitWorkTree.Helpers;
+using GitWorkTree.Services;
 using System;
 using System.Threading.Tasks;
 using System.Windows.Input;
